add: change iter impl && and new mapping
- All iterator implementation will be moved to their single files, and can stil be imported from `monad_std.iter.impl`.
- Add `map_while` and `map_windows`
