+++ docs_src/docs/Api Document/Iterator Tools Impl.md
@@ -8,82 +8,94 @@
     options:
         heading_level: 2
 
-::: monad_std.iter.builtin_like.Enumerate
-    options:
-        heading_level: 2
-
-::: monad_std.iter.builtin_like.Filter
-    options:
-        heading_level: 2
-
-::: monad_std.iter.builtin_like.Map
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.ArrayChunk
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.Chunk
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.Chain
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.FilterMap
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.FlatMap
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.Flatten
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.Fuse
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.Inspect
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.Intersperse
-    options:
-        heading_level: 2
-
-::: monad_std.iter.rust_like.IntersperseWith
-    options:
-        heading_level: 2
-
 ::: monad_std.iter.iter.OnceWith
     options:
         heading_level: 2
 
-::: monad_std.iter.rust_like.Peekable
+::: monad_std.iter.iter.Repeat
     options:
         heading_level: 2
 
-::: monad_std.iter.rust_like.Scan
+::: monad_std.iter.impl.array_chunk.ArrayChunk
     options:
         heading_level: 2
 
-::: monad_std.iter.rust_like.Skip
+::: monad_std.iter.impl.chain.Chain
     options:
         heading_level: 2
 
-::: monad_std.iter.rust_like.Take
+::: monad_std.iter.impl.chunk.Chunk
     options:
         heading_level: 2
 
-::: monad_std.iter.rust_like.TakeWhile
+::: monad_std.iter.impl.enumerate.Enumerate
     options:
         heading_level: 2
 
-::: monad_std.iter.rust_like.Zip
+::: monad_std.iter.impl.filter.Filter
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.filter_map.FilterMap
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.flat_map.FlatMap
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.flatten.Flatten
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.fuse.Fuse
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.inspect.Inspect
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.intersperse.Intersperse
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.intersperse.IntersperseWith
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.map.Map
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.map.MapWhile
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.map.MapWindows
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.peekable.Peekable
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.scan.Scan
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.skip.Skip
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.take.Take
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.take.TakeWhile
+    options:
+        heading_level: 2
+
+::: monad_std.iter.impl.zip.Zip
     options:
         heading_level: 2
