--- monad_std/iter/impl.py
@@ -1,51 +1,0 @@
-from .builtin_like import (
-    Enumerate,
-    Map,
-    Filter
-)
-
-from .rust_like import (
-    ArrayChunk,
-    Chunk,
-    Chain,
-    Scan,
-    Inspect,
-    Intersperse,
-    IntersperseWith,
-    Zip,
-    Skip,
-    Fuse,
-    FlatMap,
-    Flatten,
-    FilterMap,
-    Take,
-    TakeWhile,
-    Peekable,
-)
-
-from .iter import OnceWith, Repeat
-
-
-__all__ = [
-    "ArrayChunk",
-    "Chain",
-    "Chunk",
-    "Enumerate",
-    "Filter",
-    "FilterMap",
-    "FlatMap",
-    "Flatten",
-    "Fuse",
-    "Inspect",
-    "Intersperse",
-    "IntersperseWith",
-    "Map",
-    "OnceWith",
-    "Peekable",
-    "Repeat",
-    "Scan",
-    "Skip",
-    "Take",
-    "TakeWhile",
-    "Zip"
-]
